+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.6.0] - 2024-01-08
 
 ### Changed
 
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [builder] Updated Captive Portal to 1.3, hwclock to 1.2, dashboard to 1.3
 - [builder] Shared handling of touched files
 - [builder] Fixed ZIM auto-discovery: folder not mounted
+- [builder] Adding zim-manager (automatically with kiwix)
 - [catalog] Fixed WikiFundis missing a link to self via proxy for visualeditor
 - [catalog] Added Magoe App
 - [catalog] Added an icon for Edupi App